# Optimize diffing