--- a/src/Fabulous/Reconciler.fs
+++ b/src/Fabulous/Reconciler.fs
@@ -4,12 +4,8 @@ open Fabulous
 
 module Reconciler =
 
-    let private compareScalars (struct (key, a, b): struct (ScalarAttributeKey * obj * obj)) : ScalarAttributeComparison =
-        let data = AttributeDefinitionStore.getScalar key
-        data.CompareBoxed a b
-
     let update (canReuseView: Widget -> Widget -> bool) (prevOpt: Widget voption) (next: Widget) (node: IViewNode) : unit =
 
-        let diff = WidgetDiff.create(prevOpt, next, canReuseView, compareScalars)
+        let diff = WidgetDiff.create(prevOpt, next, canReuseView, fun key a b -> (AttributeDefinitionStore.getScalar key).CompareBoxed a b)
 
         node.ApplyDiff(&diff)