--- a/src/Fabulous/Reconciler.fs
+++ b/src/Fabulous/Reconciler.fs
@@ -6,6 +6,7 @@ module Reconciler =
 
     let update (canReuseView: Widget -> Widget -> bool) (prevOpt: Widget voption) (next: Widget) (node: IViewNode) : unit =
 
-        let diff = WidgetDiff.create(prevOpt, next, canReuseView, fun key a b -> (AttributeDefinitionStore.getScalar key).CompareBoxed a b)
+        let diff =
+            WidgetDiff.create(prevOpt, next, canReuseView, fun key a b -> (AttributeDefinitionStore.getScalar key).CompareBoxed a b)
 
         node.ApplyDiff(&diff)